chore: converted button catagory to tsx minus delete button prefab
