+++ package.json
@@ -52,7 +52,7 @@
     "dist"
   ],
   "dependencies": {
-    "@betty-blocks/component-sdk": "^1.10.2",
+    "@betty-blocks/component-sdk": "^1.13.0",
     "@date-io/date-fns": "^1.3.13",
     "@material-ui/core": "^4.9.11",
     "@material-ui/icons": "^4.11.2",
