+++ src/prefabs/actionButton.tsx
@@ -1,0 +1,11 @@
+import { Icon, prefab } from '@betty-blocks/component-sdk';
+import { actionButton } from './structures/actionButton';
+
+const attrs = {
+  icon: Icon.ButtonIcon,
+  category: 'BUTTON',
+  keywords: ['Button'],
+};
+export default prefab('Action Button (TS)', attrs, undefined, [
+  actionButton({}),
+]);
