+++ src/prefabs/button.tsx
@@ -1,0 +1,9 @@
+import { Icon, prefab } from '@betty-blocks/component-sdk';
+import { Button } from './structures/Button';
+
+const attrs = {
+  icon: Icon.ButtonIcon,
+  category: 'BUTTON',
+  keywords: ['Button'],
+};
+export default prefab('Button (TS)', attrs, undefined, [Button({})]);
