--- src/prefabs/logoutButton.js
@@ -1,295 +1,0 @@
-(() => ({
-  name: 'Logout Button',
-  icon: 'LogoutIcon',
-  category: 'BUTTON',
-  keywords: ['Form', 'input', 'logout', 'logoutbutton'],
-  beforeCreate: ({
-    prefab,
-    save,
-    close,
-    components: { Header, Content, Field, Footer, Text, EndpointSelector },
-  }) => {
-    const [value, setValue] = React.useState({});
-    const [showValidation, setShowValidation] = React.useState(false);
-
-    function serializeParameters(obj) {
-      return Object.entries(obj).map(([name, entry]) => ({
-        name,
-        value: entry.map((v) => JSON.stringify(v)),
-      }));
-    }
-
-    return (
-      <>
-        <Header onClose={close} title="Configure logout button" />
-        <Content>
-          <Field
-            label="Redirect after logout"
-            error={
-              showValidation && (
-                <Text color="#e82600">Selecting a page is required</Text>
-              )
-            }
-          >
-            <EndpointSelector size="large" value={value} onChange={setValue} />
-          </Field>
-        </Content>
-        <Footer
-          onClose={close}
-          onSave={() => {
-            if (!Object.keys(value).length) {
-              setShowValidation(true);
-              return;
-            }
-
-            const newPrefab = { ...prefab };
-            newPrefab.interactions[0].parameters = [
-              {
-                parameter: 'redirectTo',
-                pageId: value.pageId,
-                endpointId: value.id,
-                parameters: serializeParameters(value.params),
-              },
-            ];
-            save(newPrefab);
-          }}
-        />
-      </>
-    );
-  },
-  interactions: [
-    {
-      name: 'logout',
-      sourceEvent: 'Click',
-      type: 'Global',
-      ref: {
-        sourceComponentId: '#logoutButton',
-      },
-      parameters: [],
-    },
-  ],
-  structure: [
-    {
-      name: 'Button',
-      ref: {
-        id: '#logoutButton',
-      },
-      options: [
-        {
-          label: 'Toggle visibility',
-          key: 'visible',
-          value: true,
-          type: 'TOGGLE',
-          configuration: {
-            as: 'VISIBILITY',
-          },
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Button text',
-          key: 'buttonText',
-          value: ['Logout'],
-        },
-        {
-          value: false,
-          label: 'Full width',
-          key: 'fullWidth',
-          type: 'TOGGLE',
-        },
-        {
-          label: 'Icon',
-          key: 'icon',
-          value: 'None',
-          type: 'ICON',
-        },
-        {
-          value: 'small',
-          label: 'Icon size',
-          key: 'size',
-          type: 'CUSTOM',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            allowedInput: [
-              { name: 'Large', value: 'large' },
-              { name: 'Medium', value: 'medium' },
-              { name: 'Small', value: 'small' },
-            ],
-            condition: {
-              type: 'HIDE',
-              option: 'icon',
-              comparator: 'EQ',
-              value: 'None',
-            },
-          },
-        },
-        {
-          type: 'CUSTOM',
-          label: 'Icon position',
-          key: 'iconPosition',
-          value: 'start',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            condition: {
-              type: 'HIDE',
-              option: 'icon',
-              comparator: 'EQ',
-              value: 'None',
-            },
-            allowedInput: [
-              { name: 'Start', value: 'start' },
-              { name: 'End', value: 'end' },
-            ],
-          },
-        },
-        {
-          value: ['0rem', '0rem', '0rem', '0rem'],
-          label: 'Outer space',
-          key: 'outerSpacing',
-          type: 'SIZES',
-        },
-        {
-          label: 'Disabled',
-          key: 'disabled',
-          value: false,
-          type: 'TOGGLE',
-        },
-        {
-          label: 'Add Tooltip',
-          key: 'addTooltip',
-          value: false,
-          type: 'TOGGLE',
-        },
-        {
-          label: 'Toggle tooltip visibility',
-          key: 'hasVisibleTooltip',
-          value: true,
-          type: 'TOGGLE',
-          configuration: {
-            as: 'VISIBILITY',
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Tooltip Content',
-          key: 'tooltipContent',
-          value: ['Tips'],
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          label: 'Tooltip Placement',
-          key: 'tooltipPlacement',
-          value: 'bottom',
-          type: 'CUSTOM',
-          configuration: {
-            as: 'DROPDOWN',
-            dataType: 'string',
-            allowedInput: [
-              {
-                name: 'Top Start',
-                value: 'top-start',
-              },
-              {
-                name: 'Top',
-                value: 'top',
-              },
-              {
-                name: 'Top End',
-                value: 'top-end',
-              },
-              {
-                name: 'Right',
-                value: 'right',
-              },
-              {
-                name: 'Left',
-                value: 'left',
-              },
-              {
-                name: 'Botttom Start',
-                value: 'bottom-start',
-              },
-              {
-                name: 'Bottom',
-                value: 'bottom',
-              },
-              {
-                name: 'Bottom End',
-                value: 'bottom-end',
-              },
-            ],
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'COLOR',
-          label: 'Tooltip Background',
-          key: 'tooltipBackground',
-          value: 'Medium',
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'COLOR',
-          label: 'Tooltip Text',
-          key: 'tooltipText',
-          value: 'Black',
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          value: false,
-          label: 'Advanced settings',
-          key: 'advancedSettings',
-          type: 'TOGGLE',
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Test attribute',
-          key: 'dataComponentAttribute',
-          value: ['Button'],
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'advancedSettings',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-      ],
-      descendants: [],
-    },
-  ],
-}))();
