--- src/prefabs/openPage.js
@@ -1,357 +1,0 @@
-(() => ({
-  name: 'Open Page',
-  icon: 'OpenPageIcon',
-  category: 'BUTTON',
-  keywords: ['Button', 'open', 'page', 'openpage'],
-  beforeCreate: ({
-    prefab,
-    save,
-    close,
-    components: { EndpointSelector, Header, Content, Field, Footer, Text },
-  }) => {
-    const [endpoint, setEndpoint] = React.useState({
-      id: '',
-      pageId: '',
-      params: {},
-    });
-    const [showValidation, setShowValidation] = React.useState(false);
-    const isInternalEndpoint = typeof endpoint === 'object';
-    const isExternalEndpoint = typeof endpoint === 'string';
-    const isValidEndpoint =
-      (isInternalEndpoint && endpoint.id !== '') ||
-      (isExternalEndpoint && endpoint !== '');
-
-    React.useEffect(() => {
-      if (showValidation && isValidEndpoint) setShowValidation(false);
-    }, [isValidEndpoint, showValidation, setShowValidation]);
-
-    const onSaveHandler = () => {
-      if (!isValidEndpoint) {
-        setShowValidation(true);
-        return;
-      }
-      const newPrefab = { ...prefab };
-      if (isInternalEndpoint) {
-        newPrefab.structure[0].options[4].value = endpoint;
-      } else {
-        newPrefab.structure[0].options[2].value = 'external';
-        newPrefab.structure[0].options[5].value = [endpoint];
-      }
-      save(newPrefab);
-    };
-
-    return (
-      <>
-        <Header onClose={close} title="Configure open page button" />
-        <Content>
-          <Field
-            label="Link to"
-            error={
-              showValidation && <Text color="#e82600">Page is required</Text>
-            }
-          >
-            <EndpointSelector
-              value={endpoint}
-              allowExternal
-              onChange={(value) => setEndpoint(value)}
-            />
-          </Field>
-        </Content>
-        <Footer
-          onClose={close}
-          onSave={onSaveHandler}
-          onSkip={() => {
-            const newPrefab = { ...prefab };
-            save(newPrefab);
-          }}
-        />
-      </>
-    );
-  },
-  structure: [
-    {
-      name: 'Button',
-      options: [
-        {
-          label: 'Toggle visibility',
-          key: 'visible',
-          value: true,
-          type: 'TOGGLE',
-          configuration: {
-            as: 'VISIBILITY',
-          },
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Button text',
-          key: 'buttonText',
-          value: ['Open page'],
-        },
-        {
-          type: 'CUSTOM',
-          label: 'Link to',
-          key: 'linkType',
-          value: 'internal',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            allowedInput: [
-              { name: 'Internal page', value: 'internal' },
-              { name: 'External page', value: 'external' },
-            ],
-          },
-        },
-        {
-          value: '_self',
-          label: 'Open in',
-          key: 'linkTarget',
-          type: 'CUSTOM',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            allowedInput: [
-              { name: 'Current Tab', value: '_self' },
-              { name: 'New Tab', value: '_blank' },
-            ],
-          },
-        },
-        {
-          value: '',
-          label: 'Page',
-          key: 'linkTo',
-          type: 'ENDPOINT',
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'linkType',
-              comparator: 'EQ',
-              value: 'internal',
-            },
-          },
-        },
-        {
-          value: [''],
-          label: 'URL',
-          key: 'linkToExternal',
-          type: 'VARIABLE',
-          configuration: {
-            placeholder: 'Starts with https:// or http://',
-            condition: {
-              type: 'SHOW',
-              option: 'linkType',
-              comparator: 'EQ',
-              value: 'external',
-            },
-          },
-        },
-        {
-          value: false,
-          label: 'Full width',
-          key: 'fullWidth',
-          type: 'TOGGLE',
-          configuration: {
-            condition: {
-              type: 'HIDE',
-              option: 'variant',
-              comparator: 'EQ',
-              value: 'icon',
-            },
-          },
-        },
-        {
-          label: 'Icon',
-          key: 'icon',
-          value: 'None',
-          type: 'ICON',
-        },
-        {
-          value: 'small',
-          label: 'Icon size',
-          key: 'size',
-          type: 'CUSTOM',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            allowedInput: [
-              { name: 'Large', value: 'large' },
-              { name: 'Medium', value: 'medium' },
-              { name: 'Small', value: 'small' },
-            ],
-            condition: {
-              type: 'HIDE',
-              option: 'icon',
-              comparator: 'EQ',
-              value: 'None',
-            },
-          },
-        },
-        {
-          type: 'CUSTOM',
-          label: 'Icon position',
-          key: 'iconPosition',
-          value: 'start',
-          configuration: {
-            as: 'BUTTONGROUP',
-            dataType: 'string',
-            condition: {
-              type: 'HIDE',
-              option: 'icon',
-              comparator: 'EQ',
-              value: 'None',
-            },
-            allowedInput: [
-              { name: 'Start', value: 'start' },
-              { name: 'End', value: 'end' },
-            ],
-          },
-        },
-        {
-          value: ['0rem', '0rem', '0rem', '0rem'],
-          label: 'Outer space',
-          key: 'outerSpacing',
-          type: 'SIZES',
-        },
-        {
-          label: 'Disabled',
-          key: 'disabled',
-          value: false,
-          type: 'TOGGLE',
-        },
-        {
-          label: 'Add Tooltip',
-          key: 'addTooltip',
-          value: false,
-          type: 'TOGGLE',
-        },
-        {
-          label: 'Toggle tooltip visibility',
-          key: 'hasVisibleTooltip',
-          value: true,
-          type: 'TOGGLE',
-          configuration: {
-            as: 'VISIBILITY',
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Tooltip Content',
-          key: 'tooltipContent',
-          value: ['Tips'],
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          label: 'Tooltip Placement',
-          key: 'tooltipPlacement',
-          value: 'bottom',
-          type: 'CUSTOM',
-          configuration: {
-            as: 'DROPDOWN',
-            dataType: 'string',
-            allowedInput: [
-              {
-                name: 'Top Start',
-                value: 'top-start',
-              },
-              {
-                name: 'Top',
-                value: 'top',
-              },
-              {
-                name: 'Top End',
-                value: 'top-end',
-              },
-              {
-                name: 'Right',
-                value: 'right',
-              },
-              {
-                name: 'Left',
-                value: 'left',
-              },
-              {
-                name: 'Botttom Start',
-                value: 'bottom-start',
-              },
-              {
-                name: 'Bottom',
-                value: 'bottom',
-              },
-              {
-                name: 'Bottom End',
-                value: 'bottom-end',
-              },
-            ],
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'COLOR',
-          label: 'Tooltip Background',
-          key: 'tooltipBackground',
-          value: 'Medium',
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          type: 'COLOR',
-          label: 'Tooltip Text',
-          key: 'tooltipText',
-          value: 'Black',
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'addTooltip',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-        {
-          value: false,
-          label: 'Advanced settings',
-          key: 'advancedSettings',
-          type: 'TOGGLE',
-        },
-        {
-          type: 'VARIABLE',
-          label: 'Test attribute',
-          key: 'dataComponentAttribute',
-          value: ['Button'],
-          configuration: {
-            condition: {
-              type: 'SHOW',
-              option: 'advancedSettings',
-              comparator: 'EQ',
-              value: true,
-            },
-          },
-        },
-      ],
-      descendants: [],
-    },
-  ],
-}))();
