+++ src/prefabs/structures/Button/options/tooltip.ts
@@ -1,0 +1,80 @@
+import {
+  color,
+  option,
+  showIfTrue,
+  ThemeColor,
+  toggle,
+  variable,
+} from '@betty-blocks/component-sdk';
+
+export const tooltip = {
+  addTooltip: toggle('Add Tooltip', { value: false }),
+  hasVisibleTooltip: toggle('Toggle tooltip visibility', {
+    value: true,
+    configuration: {
+      as: 'VISIBILITY',
+      condition: showIfTrue('addTooltip'),
+    },
+  }),
+  tooltipContent: variable('Tooltip Content', {
+    value: ['Tips'],
+    configuration: {
+      condition: showIfTrue('addTooltip'),
+    },
+  }),
+  tooltipPlacement: option('CUSTOM', {
+    label: 'Tooltip Placement',
+    value: 'bottom',
+    configuration: {
+      as: 'DROPDOWN',
+      dataType: 'string',
+      allowedInput: [
+        {
+          name: 'Top Start',
+          value: 'top-start',
+        },
+        {
+          name: 'Top',
+          value: 'top',
+        },
+        {
+          name: 'Top End',
+          value: 'top-end',
+        },
+        {
+          name: 'Right',
+          value: 'right',
+        },
+        {
+          name: 'Left',
+          value: 'left',
+        },
+        {
+          name: 'Botttom Start',
+          value: 'bottom-start',
+        },
+        {
+          name: 'Bottom',
+          value: 'bottom',
+        },
+        {
+          name: 'Bottom End',
+          value: 'bottom-end',
+        },
+      ],
+      condition: showIfTrue('addTooltip'),
+    },
+  }),
+  tooltipBackground: color('Tooltip Background', {
+    value: ThemeColor.MEDIUM,
+    configuration: {
+      condition: showIfTrue('addTooltip'),
+    },
+  }),
+  tooltipText: color('Tooltip Text', {
+    value: ThemeColor.BLACK,
+    configuration: {
+      condition: showIfTrue('addTooltip'),
+    },
+  }),
+};
