+++ src/prefabs/structures/actionButton/options/advanced.ts
@@ -1,0 +1,13 @@
+import { showIfTrue, toggle, variable } from '@betty-blocks/component-sdk';
+
+export const advanced = {
+  advancedSettings: toggle('Advanced settings', {
+    value: false,
+  }),
+  dataComponentAttribute: variable('Test attribute', {
+    value: [''],
+    configuration: {
+      condition: showIfTrue('advancedSettings'),
+    },
+  }),
+};
